[gocui] Change cellPos to SearchPosition with XStart and XEnd
When we add support for searching the model, the length of a highlighted search
result isn't necessarily the same as the length of the search string. As an
example, consider that you are searching for a full commit hash; the view only
renders an abbreviated hash, so when the hash is found in the model, we only
want to highlight the abbreviated version.

Prepare for this by adding an XEnd field to cellPos; when searching the view,
this is always XStart + len(searchString).

Also, rename cellPos to SearchPosition and uppercase its fields to expose them;
we will use the struct in the API to set a model search function in the next
commit.